REBRANDING BOARDGAME -> RIBZ
Rebranding project from BoardGame to RIBZ
