+++ .github/pull_request_template.md
@@ -6,12 +6,12 @@ Please use labels (_on right side_) to mark properly your pull request and attac
 
 # PR Checklist
 Before submitting this PR, please make sure:
-> Don't forget to have a look to files [CONTIBUTING](https://github.com/Foacs/boardgame/blob/master/CONTRIBUTING.md) and [CODE_OF_CONDUCT](https://github.com/Foacs/boardgame/blob/master/CODE_OF_CONDUCT.md)
+> Don't forget to have a look to files [CONTIBUTING](https://github.com/Foacs/ribz/blob/master/CONTRIBUTING.md) and [CODE_OF_CONDUCT](https://github.com/Foacs/ribz/blob/master/CODE_OF_CONDUCT.md)
 
 - [ ] Your code builds clean without any errors or warnings
 - [ ] You are using approved terminology
 - [ ] You have added unit tests
-- [ ] [Sonar](https://sonarcloud.io/dashboard?id=Foacs_boardgame) quality gate is green on your branch
+- [ ] [Sonar](https://sonarcloud.io/dashboard?id=Foacs_ribz) quality gate is green on your branch
 - [ ] You have added/updated documentation (_include README_)
 
 >You can ask @Foacs/ribz-developers to help you.
