+++ LISEZMOI.md
@@ -1,12 +1,12 @@
-_[English](https://github.com/Foacs/boardgame/blob/master/README.md)_
-# BoardGame (Jeu de plateau)
+_[English](https://github.com/Foacs/ribz/blob/master/README.md)_
+# RIBZ (RIBZ: Inspired By Zombies)
 
-![GitHub top language](https://img.shields.io/github/languages/top/Foacs/boardgame)
-[![Build Status](https://travis-ci.com/Foacs/boardgame.svg?branch=master)](https://travis-ci.com/Foacs/boardgame)
-[![codecov](https://codecov.io/gh/Foacs/boardgame/branch/master/graph/badge.svg)](https://codecov.io/gh/Foacs/boardgame)
-[![Quality Gate Status](https://sonarcloud.io/api/project_badges/measure?project=Foacs_boardgame&metric=alert_status)](https://sonarcloud.io/dashboard?id=Foacs_boardgame)
+![GitHub top language](https://img.shields.io/github/languages/top/Foacs/ribz)
+[![Build Status](https://travis-ci.com/Foacs/ribz.svg?branch=master)](https://travis-ci.com/Foacs/ribz)
+[![codecov](https://codecov.io/gh/Foacs/ribz/branch/master/graph/badge.svg)](https://codecov.io/gh/Foacs/ribz)
+[![Quality Gate Status](https://sonarcloud.io/api/project_badges/measure?project=Foacs_ribz&metric=alert_status)](https://sonarcloud.io/dashboard?id=Foacs_ribz)
 
-![GitHub release (latest SemVer including pre-releases)](https://img.shields.io/github/v/release/foacs/boardgame?include_prereleases)
+![GitHub release (latest SemVer including pre-releases)](https://img.shields.io/github/v/release/foacs/ribz?include_prereleases)
 ![License](https://img.shields.io/badge/license-CeCILL-blue)
 
 
@@ -40,6 +40,6 @@ Pour le lancer:
 ```
 
 ## Licence
-> Vous pouvez accéder à la version complète de la licence [ici](https://github.com/Foacs/boardgame/blob/master/LICENCliE.md)
+> Vous pouvez accéder à la version complète de la licence [ici](https://github.com/Foacs/ribz/blob/master/LICENCliE.md)
 
 Ce projet est sous les termes de la licence __CeCILL__.
