+++ core/src/test/java/fr/foacs/ribz/core/utils/PropertiesLoaderTest.java
@@ -1,6 +1,6 @@
 /*
  * Copyright or © or Copr. Foacs
- * contributor(s): Alexis DINQUER (13/09/2020 18:11)
+ * contributor(s): Alexis DINQUER (17/09/2020 19:09)
  *
  * adinquer@yahoo.com
  *
@@ -34,26 +34,47 @@
  * knowledge of the CeCILL license and that you accept its terms.
  */
 
-package fr.foacs.boardgame.core;
+package fr.foacs.ribz.core.utils;
 
-import com.badlogic.gdx.Game;
-import fr.foacs.boardgame.core.utils.PropertiesLoader;
-import lombok.extern.slf4j.Slf4j;
+
+import org.junit.jupiter.api.Test;
+
 import java.util.Properties;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-/**
- * Entry point class
- */
-@Slf4j
-public class BoardGame extends Game {
-  @Override
-  public void create() {
-    Properties versionProperties = PropertiesLoader.loadProperties("version");
+class PropertiesLoaderTest {
 
-    log.info("Starting BoardGame version {} powered by libGDX version {} and java version {}",
-        versionProperties.getOrDefault("version", "UNKNOWN"),
-        versionProperties.getOrDefault("libgdx_version", "UNKNOWN"),
-        System.getProperty("java.version"));
+  @Test
+  void test_loadProperties_nullPropertiesName() {
+    assertThrows(IllegalArgumentException.class, () -> PropertiesLoader.loadProperties(null));
+  }
+
+  @Test
+  void test_loadProperties_emptyPropertiesName() {
+    assertThrows(IllegalArgumentException.class, () -> PropertiesLoader.loadProperties(""));
+  }
+
+  @Test
+  void test_loadProperties_blankPropertiesName() {
+    assertThrows(IllegalArgumentException.class, () -> PropertiesLoader.loadProperties("   "));
+  }
+
+  @Test
+  void test_loadProperties_unvalidPropertiesName() {
+    final Properties properties = PropertiesLoader.loadProperties("invalid");
+    assertNotNull(properties);
+    assertTrue(properties.entrySet().isEmpty());
+  }
+
+  @Test
+  void test_loadProperties_validPropertiesName() {
+    final Properties properties = PropertiesLoader.loadProperties("loader_test");
+    assertNotNull(properties);
+    assertTrue(properties.containsKey("test"));
+    assertEquals("A test value", properties.get("test"));
   }
 }
