+++ desktop/build.gradle
@@ -18,7 +18,7 @@ task createProperties(dependsOn: processResources) {
 classes.dependsOn createProperties
 
 
-project.ext.mainClassName = "fr.foacs.boardgame.desktop.DesktopLauncher"
+project.ext.mainClassName = "fr.foacs.ribz.desktop.DesktopLauncher"
 
 task run(dependsOn: classes, type: JavaExec) {
     main = project.mainClassName
